matter-thermal-mesh/dashboard/backend/requirements.txt
flask==3.1.0
flask-socketio==5.5.1
flask-cors==5.0.1
eventlet==0.37.0
requests==2.32.3
pyyaml==6.0.2
